+++ src/main/java/net/resultadofinal/micomercial/service/CompraS.java
@@ -11,7 +11,7 @@ import java.util.List;
 public interface CompraS {
 	List<Compra> listar(int start, boolean estado, String search, int length, Long cod_per, String fini, String ffin, Integer gestion,Short tipo);
 	Compra obtener(Long cod_com);
-	Boolean adicionar(Compra c, Long productos[], Integer cantidades[], BigDecimal precios[], BigDecimal descuentos[], BigDecimal subtotales[],
+	Boolean adicionar(Compra c, Long productos[], BigDecimal cantidades[], BigDecimal precios[], BigDecimal descuentos[], BigDecimal subtotales[],
 					  BigDecimal totales[], Short tipos[]);
 	Boolean eliminar(Long cod_com,Long user);
 	Compra obtenerPorArqueoCaja(Long codArqcaj, Integer codDet);
